+++ apps/web/src/app/(app)/claw/kilo-chat/hooks/useMessages.ts
@@ -1,10 +1,12 @@
-import { useMutation, useQueryClient, useInfiniteQuery } from '@tanstack/react-query';
+export { useMessages, useSendMessage } from '@kilocode/kilo-chat-hooks';
+export type { SendMessageVariables } from '@kilocode/kilo-chat-hooks';
+
+import { useMutation, useQueryClient } from '@tanstack/react-query';
 import type { InfiniteData } from '@tanstack/react-query';
 import type { KiloChatClient } from '@kilocode/kilo-chat';
 import type {
   Message,
   ReactionSummary,
-  CreateMessageRequest,
   EditMessageRequest,
   MessageCreatedEvent,
   MessageUpdatedEvent,
@@ -18,8 +20,6 @@ import type {
 import { useEffect } from 'react';
 import { kiloclawConversationContext } from '@kilocode/event-service';
 import { toast } from 'sonner';
-
-const PAGE_SIZE = 50;
 
 function applyReactionAdded(
   reactions: ReactionSummary[],
@@ -106,78 +106,6 @@ function findMessageInCache(
     if (match) return match;
   }
   return undefined;
-}
-
-export function useMessages(client: KiloChatClient, conversationId: string | null) {
-  return useInfiniteQuery({
-    queryKey: ['kilo-chat', 'messages', conversationId],
-    queryFn: async ({ pageParam }) => {
-      return client.listMessages(conversationId ?? '', { before: pageParam, limit: PAGE_SIZE });
-    },
-    initialPageParam: undefined as string | undefined,
-    getNextPageParam: lastPage => {
-      if (lastPage.length < PAGE_SIZE) return undefined;
-      return lastPage[lastPage.length - 1]?.id;
-    },
-    enabled: !!conversationId,
-    select: data => ({
-      ...data,
-      messages: data.pages.flatMap(p => p).reverse(),
-    }),
-  });
-}
-
-export type SendMessageVariables = CreateMessageRequest & { clientId: string };
-
-export function useSendMessage(
-  client: KiloChatClient,
-  conversationId: string | null,
-  currentUserId: string
-) {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: (req: SendMessageVariables) => client.sendMessage(req),
-    onMutate: async (variables: SendMessageVariables) => {
-      if (!conversationId) return;
-      const queryKey = ['kilo-chat', 'messages', conversationId];
-      await queryClient.cancelQueries({ queryKey });
-      const pendingId = `pending-${variables.clientId}`;
-      const optimisticMessage: Message = {
-        id: pendingId,
-        senderId: currentUserId,
-        content: variables.content,
-        inReplyToMessageId: variables.inReplyToMessageId ?? null,
-        updatedAt: null,
-        clientUpdatedAt: null,
-        deleted: false,
-        deliveryFailed: false,
-        reactions: [],
-      };
-      queryClient.setQueryData<InfiniteData<Message[]>>(queryKey, old => {
-        if (!old) return old;
-        const firstPage = old.pages[0] ?? [];
-        return { ...old, pages: [[optimisticMessage, ...firstPage], ...old.pages.slice(1)] };
-      });
-      return { queryKey, pendingId };
-    },
-    onSuccess: (response, _variables, context) => {
-      if (!context) return;
-      const { queryKey, pendingId } = context;
-      queryClient.setQueryData<InfiniteData<Message[]>>(queryKey, old => {
-        if (!old) return old;
-        return {
-          ...old,
-          pages: old.pages.map(page =>
-            page.map(msg => (msg.id === pendingId ? { ...msg, id: response.messageId } : msg))
-          ),
-        };
-      });
-    },
-    onError: (_err, _variables, context) => {
-      if (!context) return;
-      removeMessageFromCache(queryClient, context.queryKey, context.pendingId);
-    },
-  });
 }
 
 export function useEditMessage(client: KiloChatClient, conversationId: string | null) {
