+++ packages/kilo-chat-hooks/src/index.ts
@@ -1,2 +1,3 @@
 export * from './context';
 export * from './use-conversations';
+export * from './use-messages';
